move statemachine to its own directory
